# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Course An Ninh Mang API",
    "version": "1.0.0",
    "description": "API documentation for Course An Ninh Mang backend",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Development server"
    },
    {
      "url": "https://course-an-ninh-mang-backend.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "error_code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error_code": {
            "type": "integer",
            "example": 40001
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/courses": {
      "get": {
        "summary": "Get all courses",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of courses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/management": {
      "get": {
        "summary": "Get courses for management (admin)",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Courses management data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "summary": "Get course detail with progress",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Course ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Course detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not enrolled in course",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}/content": {
      "get": {
        "summary": "Get course content (lesson content)",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Lesson ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Lesson content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}/enroll": {
      "post": {
        "summary": "Enroll in a course",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Course ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully enrolled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "409": {
            "description": "Already enrolled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/upload": {
      "post": {
        "summary": "Upload course (admin)",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Course uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payment": {
      "post": {
        "summary": "Create payment request",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "ho_ten",
                  "email",
                  "ten_goi",
                  "so_tien",
                  "proof_image"
                ],
                "properties": {
                  "ho_ten": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "so_dien_thoai": {
                    "type": "string"
                  },
                  "phuong_thuc_thanh_toan": {
                    "type": "string"
                  },
                  "ten_goi": {
                    "type": "string",
                    "enum": [
                      "Gói Cơ Bản",
                      "Gói Nâng Cao",
                      "Gói Năm"
                    ]
                  },
                  "so_tien": {
                    "type": "number"
                  },
                  "proof_image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all payments (filtered by user or admin)",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "trang_thai",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "completed",
                "rejected"
              ]
            }
          },
          {
            "in": "query",
            "name": "ten_goi",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/{id}": {
      "get": {
        "summary": "Get payment by ID",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/{id}/approve": {
      "put": {
        "summary": "Approve payment (admin only)",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/{id}/reject": {
      "put": {
        "summary": "Reject payment (admin only)",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/confirm": {
      "post": {
        "summary": "Confirm payment (backward compatibility)",
        "tags": [
          "Payment"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload/upload": {
      "post": {
        "summary": "Upload file directly (server-side)",
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "folder": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}